requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
datasets>=2.12.0

# Computer Vision & Image Processing
Pillow>=9.5.0
opencv-python>=4.8.0

# OCR (Optional - for advanced dataset with text extraction)
pytesseract>=0.3.10
easyocr>=1.7.0

# Data handling
numpy>=1.24.0
pandas>=2.0.0
PyYAML>=6.0

# Logging & Visualization
wandb>=0.15.0
rich>=13.4.0
tqdm>=4.65.0
tensorboard>=2.13.0

# Utilities
python-dateutil>=2.8.2
scikit-learn>=1.2.0
